options: fix window-scale property
A recent change merged the window-scaler option and property, but forgot
that the option is float for some reason, while the property uses
double. This led to undefined behavior. Fix it by changing the option
to double too.

+++ options/options.h
@@ -34,7 +34,7 @@ typedef struct mp_vo_opts {
     struct m_geometry autofit;
     struct m_geometry autofit_larger;
     struct m_geometry autofit_smaller;
-    float window_scale;
+    double window_scale;
 
     int keepaspect;
     int keepaspect_window;
